backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
pytz==2023.3
apscheduler==3.10.4
python-multipart==0.0.6
pydantic==2.4.2
alembic==1.12.0
httpx
pinecone>=6.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic-extra-types==2.1.0
pydantic-settings==2.0.3
pydantic_core==2.10.1
python-jose==3.3.0
PyYAML==6.0.1
cryptography==41.0.5
email-validator